ref(landing): Move Kalpa banner above header in App layout
Extract KalpaBanner into its own component and render it above
HeaderBar in App.tsx so the announcement banner sits at the very
top of the viewport instead of below the header.
